relocate cache factory func
